Phase One Capstone Project

+++ src/main/java/lab1/Account.java
@@ -27,7 +27,7 @@ public abstract class Account {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
